The laboratory page is divided into components. Prepared projects have been added to the homepage. JSON (blakfyProjectData.json) file was created for blakfy projects.
