+++ app/(Pages)/laboratory/page.tsx
@@ -1,13 +1,13 @@
 import { Container } from "@radix-ui/themes";
 import ImageSection from "./Components/ImageSection";
 import TextSection from "./Components/TextSection";
-import CardItem from "./Components/CardItem";
+import ProjectContainer from "./Components/ProjectsContainer";
 export default function Home() {
   return (
     <Container size={"2"}>
       <ImageSection />
       <TextSection />
-      <CardItem />
+      <ProjectContainer />
     </Container>
   );
 }
